Enhance BLE transfer functionality with delegate methods for powered on and connection status
