+++ Sources/MdocDataTransfer18013/BLETransfer/ServiceCharacteristics.swift
@@ -19,21 +19,6 @@ import Foundation
 import CoreBluetooth
 import SwiftCBOR
 
-/// The enum BleTransferMode defines the two roles in the communication, which can be a server or a client.
-///
-/// The four static variables are used to signal the start and the end of the communication. This is done by sending the bytes 0x01 and 0x02 for the start and end of the communication, respectively. For the start and end of the data transmission, the bytes 0x01 and 0x00 are used.
-public enum BleTransferMode: Sendable {
-	case server
-	case client
-	// signals for coordination
-	static let START_REQUEST: [UInt8] = [0x01]
-	static let END_REQUEST: [UInt8] = [0x02]
-	static let START_DATA: [UInt8] = [0x01]
-	static let END_DATA: [UInt8] = [0x00]
-	public static let BASE_UUID_SUFFIX_SERVICE = "-0000-1000-8000-00805F9B34FB"
-	public static let QRHandover = CBOR.null
-}
-
 extension MdocGattServer {
 	/// mdoc service characteristic definitions (mdoc is the GATT server)
 	public enum MdocServiceCharacteristic: String, CustomStringConvertible, Sendable {
